+++ site_builder/name_conversion.py
@@ -31,10 +31,8 @@ class NameConverter:
         self.__category_settings = AliasStore('category', settings.category)
         self.__forum_alias = AliasStore('forum', settings.forum)
         self.__word_alias = AliasStore('word', settings.word)
-        self.__channel_alias = {
-            source_dir / Path(channel['path']) if channel.get('path') else source_dir: channel['name']
-            for channel in DiscordChannelID.CHANNEL_LOOKUP.values()
-        }
+        self.__channel_alias = {source_dir / channel['path']: channel['name'] for channel in
+                                DiscordChannelID.CHANNEL_LOOKUP.values() if channel['path'] is not None}
         self.__extra_channel_alias = AliasStore('extra-channel-alias', settings.extra_channel)
 
     def channel(self, source_file: Path):